pyproject.toml
[project]
name = "city-growth-ai-agent"
version = "0.1.0"
description = "AI agent for urban economics analysis with QCEW data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core data processing
    "pandas>=2.0.0",
    "numpy",
    "plotly",
    "matplotlib",
    "adjustText",
    # Database
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    # LangChain/LangGraph
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    # FastAPI backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    # Checkpointer for LangGraph persistence (PostgreSQL)
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary,pool]>=3.2.0",
    # Utilities
    "python-dotenv>=1.0.0",
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ipykernel",
    "playwright>=1.57.0",
]
